Updated enpoints to use mongodb data instead, exported mongodb model to external file, added options when seeding the database
